+++ packages/expo-router/src/layouts/stack-utils/shared.ts
@@ -1,62 +1,16 @@
 import type { NativeStackHeaderItemButton } from '@react-navigation/native-stack';
 import { Children, type ReactNode } from 'react';
-import {
-  StyleSheet,
-  type ColorValue,
-  type ImageSourcePropType,
-  type StyleProp,
-  type TextStyle,
-} from 'react-native';
+import { type ColorValue, type ImageSourcePropType, type StyleProp } from 'react-native';
 import type { SFSymbol } from 'sf-symbols-typescript';
 
 import { StackHeaderBadge, StackHeaderIcon, StackHeaderLabel } from './common-primitives';
 import { getFirstChildOfType } from '../../utils/children';
+import { convertTextStyleToRNTextStyle, type BasicTextStyle } from '../../utils/font';
 
 export interface StackHeaderItemSharedProps {
-  /**
-   * There are two ways to specify the content of the header item:
-   *
-   * @example
-   * ```tsx
-   * import { Stack } from 'expo-router';
-   *
-   * ...
-   * <Stack.Header.Button>As text passed as children</Stack.Header.Button>
-   * ```
-   *
-   * @example
-   * ```tsx
-   * import { Stack } from 'expo-router';
-   *
-   * ...
-   * <Stack.Header.Button>
-   *   <Stack.Header.Icon sf="star.fill" />
-   *   <Stack.Header.Label>As components</Stack.Header.Label>
-   *   <Stack.Header.Badge>3</Stack.Header.Badge>
-   * </Stack.Header.Button>
-   * ```
-   *
-   * **Note**: When icon is used, the label will not be shown and will be used for accessibility purposes only.
-   */
   children?: ReactNode;
-  /**
-   * Style for the label of the header item.
-   */
-  style?: StyleProp<
-    Pick<TextStyle, 'fontFamily' | 'fontSize' | 'fontWeight' | 'color'> & {
-      /**
-       * When set to 'transparent', the button will have no background color.
-       *
-       * @platform iOS 26+
-       */
-      backgroundColor?: 'transparent';
-    }
-  >;
-  /**
-   * Whether to separate the background of this item from other header items.
-   *
-   * @default false
-   */
+  style?: StyleProp<BasicTextStyle>;
+  hidesSharedBackground?: boolean;
   separateBackground?: boolean;
   accessibilityLabel?: string;
   accessibilityHint?: string;
@@ -124,11 +78,8 @@ export function convertStackHeaderSharedPropsToRNSharedHeaderItem(
     sharesBackground: !separateBackground,
   };
   if (style) {
-    const { backgroundColor, ...convertedStyle } = convertTextStyleToRNTextStyle(style) ?? {};
+    const convertedStyle = convertTextStyleToRNTextStyle(style) ?? {};
     item.labelStyle = convertedStyle;
-    item.hidesSharedBackground = backgroundColor === 'transparent';
-  } else {
-    item.hidesSharedBackground = false;
   }
   if (badgeComponent) {
     item.badge = {
@@ -143,31 +94,4 @@ export function convertStackHeaderSharedPropsToRNSharedHeaderItem(
     item.icon = rnsIcon;
   }
   return item;
-}
-
-type NumericFontWeight = 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900;
-
-type ConvertedFontWeightType = Exclude<TextStyle['fontWeight'], number> | `${NumericFontWeight}`;
-
-function convertTextStyleToRNTextStyle<BaseStyleType extends Pick<TextStyle, 'fontWeight'>>(
-  style: StyleProp<BaseStyleType | undefined>
-):
-  | (Omit<BaseStyleType, 'fontWeight'> & {
-      fontWeight?: ConvertedFontWeightType;
-    })
-  | undefined {
-  const flattenedStyle = StyleSheet.flatten(style) as BaseStyleType | undefined;
-  if (!flattenedStyle) {
-    return undefined;
-  }
-  if ('fontWeight' in flattenedStyle) {
-    return {
-      ...flattenedStyle,
-      fontWeight:
-        typeof flattenedStyle.fontWeight === 'number'
-          ? (String(flattenedStyle.fontWeight) as `${NumericFontWeight}`)
-          : flattenedStyle.fontWeight,
-    };
-  }
-  return flattenedStyle as Omit<BaseStyleType, 'fontWeight'>;
 }